pyproject.toml
[tool.poetry]
name = "keller_pet"
version = "0.6.0"
description = "API to help PetShop owners store data from the pets."
authors = ["Jefte Keller <kellergnr@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2"
djangorestframework = "^3.12.4"

[tool.poetry.dev-dependencies]
pylint = "^2.8.2"
yapf = "^0.31.0"
autopep8 = "^1.5.7"
ipdb = "^0.13.7"
pylint-django = "^2.4.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
